+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 
 - Added `signal_handler_timeout_seconds` parameter, which controls the timeout duration for the signal handler. Only
   available on Linux platforms.
+- Avoid calling `std::this_thread::sleep_for()` in the signal handler when waiting for the log to be flushed.
 
 ## v4.1.0